moved filter/sort selector view into its own file
